// src/app/models/basic-details.ts
export class BasicDetails{
    maritalStatus: string;
    options: boolean;
    fullName:string;
    dob:any;
    gender:string;
    motherName:string;
    fatherName:string;
    single:string;
    married:string;
    occupation: string;
    addressLine1: any;
    addressLine2: any;
    state: string;
    pinCode: any;
    city: any;
    addressType: any;
    incomeSlab: any;
    pep: any;

    constructor(){

    }
}